Added authorization for post editing and deleting.

+++ backend/routes/posts.js
@@ -31,7 +31,8 @@ router.post("", checkAuth, multer({storage: storage}).single("image"), (req, res
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + "/images/" + req.file.filename
+    imagePath: url + "/images/" + req.file.filename,
+    creator: req.userData.userId
   });
 
   post.save().then((storedData) => {
@@ -87,22 +88,30 @@ router.put("/:id", checkAuth, multer({storage: storage}).single("image"), (req, 
     _id: req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath: imagePath
+    imagePath: imagePath,
+    creator: req.userData.userId
   });
 
-  Post.updateOne({_id: req.params.id}, post).then((result) => {
-    res.status(200).json({
-      message: "Post edited succesfully."
-    });
+  Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post).then((result) => {
+    if(result.nModified > 0) {
+      res.status(200).json({ message: "Post edited successfully."});
+    } else {
+      res.status(401).json({ message: "Not authorized for editing the post."})
+    }
   });
 })
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({_id: req.params.id}).then((result) => {
-    console.log(result);
-    res.status(200).json({
-      message: "Post Deleted"
-    });
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then((result) => {
+    if(result.n > 0) {
+      res.status(200).json({
+        message: "Post Deleted successfully"
+      });
+    } else {
+      res.status(401).json({
+        message: "Not authorized for deleting the post."
+      })
+    }
   });
 })
